fix Bettersearch/extended tag button
when adding the toolbar this add-on checks if extended tag dialog has added an
entry to the browser menu. So the function make_and_add_toolbar from this
add-on must run after the hooked functions from the other add-ons have run.
So I may not use a gui_hook that runs early during the browser setup but
instead I must one that is run at the end of the initialization.