qtawesome 0.5 -> 1.1.1
the old qtawesome code caused a deprecation warning that a
function used might break in future python versions.

So I downloaded qtawesome 1.1.1 from 2021-11-29 from
https://github.com/spyder-ide/qtawesome/releases

From this release I removed all files not used (to minimize
the size of the fastbar add-on).

I modified the remaining files:
- I adjusted the imports so that qt stuff is imported from
  aqt.qt.
- I changed to the new-style PyQt/qt6 scoped enums
- I commented out the parts that would load the files I
  removed.

Since this newer qtawesome package is python3 only we no
longer need six (which was a Python 2 and 3 compatibility
library).