allow actions from other actions shown (if set in the user config)
e.g. the addon "Browser Maximize/Hide Table/Editor" from
https://ankiweb.net/shared/info/1819291495 adds a QAction
named maximize_addon__action_editor_only to the browser
instance.

With this commit you can add a button to the fastbar for this
action by adding the following to the add-on config:

[
  "maximize_addon__action_editor_only",
  "fa.pause-circle"
]

This makes the fastbar extensible by other add-ons and users can
still set the buttons they want to see. It's not user-friendly,
but flexible.